pyproject.toml
[project]
name = "wan-automation-examples"
version = "0.1.0"
description = "Python and Terraform examples for interacting with Cisco Catalyst SD-WAN Manager and the Meraki Dashboard APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "datetime>=5.5",
    "dotenv>=0.9.9",
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "simple-term-menu>=1.6.6",
    "tabulate>=0.9.0",
]
